refactor(util): split klaw-util modules
Separate path helpers, environment types, and command PATH augmentation
into focused modules so the crate root stays small while preserving the
existing public re-export surface for downstream callers.

Made-with: Cursor
